refactor VIPER downloads
